\name{ee1}
\alias{ee1}
\title{TODO...}
\usage{ee1(x, y, x0=x, y0=y, type=c("none", "dec", "inc"), m=20, lqi=0.05,
    hqi=0.95, sym=FALSE, plot=TRUE, flag=FALSE, dg=15, logtran=FALSE)}
\description{TODO}
\value{TODO}
\arguments{\item{x}{TODO}
\item{y}{TODO}
\item{x0}{TODO}
\item{y0}{TODO}
\item{type}{one of "none", "dec" or "inc"}
\item{m}{defaults}
\item{lqi}{TODO}
\item{hqi}{TODO}
\item{sym}{TODO}
\item{plot}{TODO}
\item{flag}{TODO}
\item{dg}{TODO}
\item{logtran}{TODO}}
