\name{f.cut0}
\alias{f.cut0}
\title{function to define the cutoffs to be used for each individual gene set statistic;...}
\usage{f.cut0(w0, w)}
\description{function to define the cutoffs to be used for each individual gene set statistic;}
\value{vector of p values for each gene set}
\arguments{\item{w0}{vector of observed gene set statistics
TODO rename to observedGenesetStats}
\item{w}{matrix of gene set statistics based on the permutation procedure
TODO rename to permutationGenesetStats}}
