\name{f.GOInputMLP}
\alias{f.GOInputMLP}
\title{Prepare the input geneSet for the f...}
\usage{f.GOInputMLP(goInFeatureNames)}
\description{Prepare the input geneSet for the f.MLP function}
\value{data structure ready to be used as geneSet argument for the f.MLP
function}
\arguments{\item{goInFeatureNames}{output from f.GOAnnotation}}
