\name{f.permtwo}
\alias{f.permtwo}
\title{Calculate all unique permutations for 1:n, with k = 2 in each group...}
\usage{f.permtwo(n, k)}
\description{Calculate all unique permutations for 1:n, with k = 2 in each group
(e.g. wild-type vs. knock-out) ; used for column permutations only}
\arguments{\item{n}{}
\item{k}{}}
