% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatKEGGDescription}
\alias{formatKEGGDescription}
\title{Format KEGG descriptions}
\usage{
formatKEGGDescription(descriptions, species = "Mouse")
}
\arguments{
\item{descriptions}{character vector with descriptions of 
KEGG pathways}

\item{species}{character vector of length one indicating the species, one of
'Mouse', 'Human', 'Rat', 'Dog' or 'Rhesus'; defaults to 'Mouse'.}
}
\value{
formatted descriptions
}
\description{
This includes the removal of the species
}
\keyword{internal}
