\name{goInputMLP}
\alias{goInputMLP}
\title{Prepare the input geneSet for the MLP function...}
\usage{goInputMLP(goInFeatureNames)}
\description{Prepare the input geneSet for the MLP function}
\value{data structure ready to be used as geneSet argument for the MLP
function}
\arguments{\item{goInFeatureNames}{output from goAnnotation}}
