\name{mlpStatistic}
\alias{mlpStatistic}
\title{This function calculates the mean of the gene-statistic y[, 2] (the MLP statistic)...}
\usage{mlpStatistic(geneSet, genePValue, mappingFunctionOutput)}
\description{This function calculates the mean of the gene-statistic y[, 2] (the MLP statistic)
for each unique gene-set in x[, 1].}
\value{data frame with two columns; first column contains the number of genes
in the gene set; the second column contains the MLP statistic (u) for the gene set.}
\arguments{\item{geneSet}{is a data frame of with two columns representing the gene-sets and gene-names
respectively; a geneSet data frame is typically retrieved from a BioConductor annotation
package.}
\item{genePValue}{is a data frame with a column of gene-names and a column containing a 
gene-statistic for each gene name; the genePValue data frame is based on the data under
study and the statistics (or p values) can e.g. be generated by limma.}
\item{mappingFunctionOutput}{is the mapping of the gene-names in genePValue to those in geneSet;
the mapping is the output of a mapping function.}}
