\name{plotQuantileCurves}
\alias{plotQuantileCurves}
\title{Plot Quantile Curves for an MLP Run...}
\usage{plotQuantileCurves(x0, y0, hqi, xtp, qi, lqi, sym=TRUE)}
\description{Plot Quantile Curves for an MLP Run}
\value{no return value; a quantile curve plot is drawn to the current device}
\arguments{\item{x0}{TODO}
\item{y0}{TODO}
\item{hqi}{TODO}
\item{xtp}{TODO}
\item{qi}{TODO}
\item{lqi}{TODO}
\item{sym}{TODO defaults to TRUE}}
