\name{pp2g}
\alias{pp2g}
\title{Function to convert probeset datasets to gene datasets...}
\usage{pp2g(xp, yp)}
\description{Function to convert probeset datasets to gene datasets
(deprecated) 
TODO look for BioConductor based equivalent}
\value{TODO fill out}
\arguments{\item{xp}{input table of gene-sets, probe.ids, and gene.ids, in that order.}
\item{yp}{input table of probe.ids and original and permuted probeset-statistics.}}
