\name{rssp}
\alias{rssp}
\title{computes the pooled standard deviation for two groups...}
\usage{rssp(x1, x2)}
\description{computes the pooled standard deviation for two groups}
\arguments{\item{x1}{matrix corresponding to the first of two groups}
\item{x2}{matrix corresponding to the second of two groups}}
