\name{addGeneSetDescription}
\alias{addGeneSetDescription}
\title{TODO (low priority currently) add print method for MLP objects...}
\usage{addGeneSetDescription(object, geneSets, ...)}
\description{TODO (low priority currently) add print method for MLP objects
Summary function for MLP objects; this function combines
the MLP results with the biological description of the
gene sets}
\value{TODO}
\arguments{\item{object}{object of class 'MLP' as produced by the 'MLP' function}
\item{geneSets}{list of gene sets; each gene set (component) contains
the member genes as a character vector. Typically, the output of goAnnotation
is used.}
\item{...}{further arguments; currently none are used}}
