\name{ctpval1}
\alias{ctpval1}
\title{This function interpolates between the smoothed quantile curves to determine the p value for a given geneset...}
\usage{ctpval1(x.ct, q.cutoff)}
\description{This function interpolates between the smoothed quantile curves to determine the p value for a given geneset}
\value{actual p values for each gene set}
\arguments{\item{x.ct}{output of quantileCurves, i.e. the curves of the contours for the different quantiles}
\item{q.cutoff}{critical values}}
