\name{plotGOgraph}
\alias{plotGOgraph}
\title{Graphical Representation of GO Based MLP Results...}
\usage{plotGOgraph(object, ontology, annotation, nRow=5, mainTitle="GO graph")}
\description{Graphical Representation of GO Based MLP Results}
\value{GO graph is plotted to the current device}
\arguments{\item{object}{object of class MLP (as produced by the MLP function)}
\item{ontology}{ontology}
\item{annotation}{annotation}
\item{nRow}{nRow}
\item{mainTitle}{main title of the graph; defaults to 'GO graph'}}
