\name{plotGOgraph}
\alias{plotGOgraph}
\title{Graphical Representation of GO Based MLP Results...}
\usage{plotGOgraph(object, nRow=5, main="GO graph")}
\description{Graphical Representation of GO Based MLP Results}
\value{GO graph is plotted to the current device}
\arguments{\item{object}{object of class MLP (as produced by the MLP function)}
\item{nRow}{number of GO IDs for which to produce the plot}
\item{main}{main title of the graph; defaults to 'GO graph'}}
