\name{summary.MLP}
\alias{summary.MLP}
\title{TODO fix S3 export...}
\usage{\method{summary}{MLP} (object, goInFeatureNames, ...)}
\description{TODO fix S3 export
TODO 
TODO (low priority currently) add print method for MLP objects
Summary function for MLP objects; this function combines
the MLP results with the biological description of the
gene sets}
\value{TODO}
\arguments{\item{object}{object of class 'MLP' as produced by the 'MLP' function}
\item{goInFeatureNames}{}
\item{...}{further arguments; currently none are used}}
